Moving expiration and token type properties to the parent class
