+++ security-jwt/src/main/java/io/micronaut/security/token/jwt/render/AccessRefreshToken.java
@@ -15,6 +15,7 @@
  */
 package io.micronaut.security.token.jwt.render;
 
+import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 /**
@@ -31,6 +32,13 @@ public class AccessRefreshToken {
     @JsonProperty("refresh_token")
     private String refreshToken;
 
+    @JsonProperty("token_type")
+    private String tokenType;
+
+    @JsonProperty("expires_in")
+    @JsonInclude(JsonInclude.Include.NON_NULL)
+    private Integer expiresIn;
+
     /**
      * Necessary for JSON serialization.
      */
@@ -40,14 +48,30 @@ public class AccessRefreshToken {
      *
      * @param accessToken JWT token
      * @param refreshToken JWT token
+     * @param tokenType Type of token
      */
-    public AccessRefreshToken(String accessToken, String refreshToken) {
+    public AccessRefreshToken(String accessToken, String refreshToken, String tokenType) {
         this.accessToken = accessToken;
         this.refreshToken = refreshToken;
+        this.tokenType = tokenType;
     }
 
     /**
-     * accesToken getter.
+     *
+     * @param accessToken JWT token
+     * @param refreshToken JWT token
+     * @param tokenType Type of token
+     * @param expiresIn Seconds until token expiration
+     */
+    public AccessRefreshToken(String accessToken, String refreshToken, String tokenType, Integer expiresIn) {
+        this.accessToken = accessToken;
+        this.refreshToken = refreshToken;
+        this.tokenType = tokenType;
+        this.expiresIn = expiresIn;
+    }
+
+    /**
+     * accessToken getter.
      * @return The access token
      */
     public String getAccessToken() {
@@ -55,10 +79,26 @@ public class AccessRefreshToken {
     }
 
     /**
-     * accesToken setter.
+     * refreshToken getter.
      * @return The refresh token
      */
     public String getRefreshToken() {
         return refreshToken;
+    }
+
+    /**
+     * token type getter.
+     * @return TokenType e.g. Bearer
+     */
+    public String getTokenType() {
+        return tokenType;
+    }
+
+    /**
+     * token type getter.
+     * @return expiration time
+     */
+    public Integer getExpiresIn() {
+        return expiresIn;
     }
 }
