+++ security-jwt/src/main/java/io/micronaut/security/token/jwt/render/BearerAccessRefreshToken.java
@@ -15,9 +15,6 @@
  */
 package io.micronaut.security.token.jwt.render;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
-import io.micronaut.http.HttpHeaderValues;
-
 import java.util.Collection;
 import java.util.List;
 
@@ -32,12 +29,6 @@ public class BearerAccessRefreshToken extends AccessRefreshToken {
     private String username;
     private Collection<String> roles;
 
-    @JsonProperty("expires_in")
-    private Integer expiresIn;
-
-    @JsonProperty("token_type")
-    private String tokenType = HttpHeaderValues.AUTHORIZATION_PREFIX_BEARER;
-
     /**
      * Necessary for JSON serialization.
      */
@@ -47,7 +38,7 @@ public class BearerAccessRefreshToken extends AccessRefreshToken {
      *
      * @param username a string e.g. admin
      * @param roles Collection of Strings e.g. ( [ROLE_USER, ROLE_ADMIN] )
-     * @param expiresIn Acccess Token expiration
+     * @param expiresIn Access Token expiration
      * @param accessToken JWT token
      * @param refreshToken  JWT token
      */
@@ -55,11 +46,12 @@ public class BearerAccessRefreshToken extends AccessRefreshToken {
                                     Collection<String> roles,
                                     Integer expiresIn,
                                     String accessToken,
-                                    String refreshToken) {
-        super(accessToken, refreshToken);
+                                    String refreshToken,
+                                    String tokenType
+    ) {
+        super(accessToken, refreshToken, tokenType, expiresIn);
         this.username = username;
         this.roles = roles;
-        this.expiresIn = expiresIn;
     }
 
     /**
@@ -92,37 +84,5 @@ public class BearerAccessRefreshToken extends AccessRefreshToken {
      */
     public void setRoles(List<String> roles) {
         this.roles = roles;
-    }
-
-    /**
-     *
-     * @return TokenType e.g. Bearer
-     */
-    public String getTokenType() {
-        return tokenType;
-    }
-
-    /**
-     * token type setter.
-     * @param tokenType e.g. Bearer
-     */
-    public void setTokenType(String tokenType) {
-        this.tokenType = tokenType;
-    }
-
-    /**
-     *
-     * @return expiration time
-     */
-    public Integer getExpiresIn() {
-        return expiresIn;
-    }
-
-    /**
-     *
-     * @param expiresIn expiration time
-     */
-    public void setExpiresIn(Integer expiresIn) {
-        this.expiresIn = expiresIn;
     }
 }
