+++ src/main/docs/guide/appendix/breaks.adoc
@@ -4,6 +4,9 @@ This section will document breaking changes that may happen during milestone or 
 
 * link:{api}/io/micronaut/session/http/HttpSessionConfiguration.html[HttpSessionConfiguration] default value for `cookiePath` is `/`.
 
+* link:{api}/io/micronaut/security/token/jwt/render/AccessRefreshToken.html[AccessRefreshToken] includes `tokenType` and `expiresIn` to comply with refresh token responses specified in link:https://tools.ietf.org/html/rfc6749#section-4.1.4.html[RFC 6749].
+These properties are serialized respectively as `token_type` and `expires_in` in the refresh token response.
+
 ==== 1.0.0.RC3
 
 * All Micronaut modules have been renamed to include `micronaut-` prefix to make it easier to manage dependencies. If you are upgrading rename all references modules. Example `bom` -> `micronaut-bom`, `inject` -> `micronaut-inject` etc.
